+++ engine/src/test/java/net/jqwik/api/ArrayArbitraryTests.java
@@ -1,0 +1,132 @@
+package net.jqwik.api;
+
+import java.util.*;
+import java.util.stream.*;
+
+import net.jqwik.api.arbitraries.*;
+import net.jqwik.api.constraints.*;
+
+import static org.assertj.core.api.Assertions.*;
+
+import static net.jqwik.api.ArbitraryTestHelper.*;
+import static net.jqwik.testing.ShrinkingSupport.*;
+import static net.jqwik.testing.TestingSupport.*;
+
+@Group
+class ArrayArbitraryTests {
+
+	@Example
+	void array() {
+		Arbitrary<Integer> integerArbitrary = Arbitraries.integers().between(1, 10);
+		ArrayArbitrary<Integer, Integer[]> arrayArbitrary = integerArbitrary.array(Integer[].class).ofMinSize(2).ofMaxSize(5);
+
+		RandomGenerator<Integer[]> generator = arrayArbitrary.generator(1);
+
+		assertAllGenerated(generator, array -> {
+			assertThat(array.length).isBetween(2, 5);
+			assertThat(array).isSubsetOf(1, 2, 3, 4, 5, 6, 7, 8, 9, 10);
+		});
+	}
+
+	@Example
+	void reduceArray() {
+		ArrayArbitrary<Integer, Integer[]> arrayArbitrary =
+				Arbitraries.integers().between(1, 5).array(Integer[].class).ofMinSize(1).ofMaxSize(10);
+
+		Arbitrary<Integer> integerArbitrary = arrayArbitrary.reduce(0, Integer::sum);
+
+		RandomGenerator<Integer> generator = integerArbitrary.generator(1000);
+
+		assertAllGenerated(generator, sum -> {
+			assertThat(sum).isBetween(1, 50);
+		});
+
+		assertAtLeastOneGenerated(generator, sum -> sum == 1);
+		assertAtLeastOneGenerated(generator, sum -> sum > 30);
+	}
+
+	@Example
+	void arrayOfPrimitiveType(@ForAll Random random) {
+		Arbitrary<Integer> integerArbitrary = Arbitraries.integers().between(1, 10);
+		ArrayArbitrary<Integer, int[]> arrayArbitrary = integerArbitrary.array(int[].class).ofMinSize(0).ofMaxSize(5);
+
+		RandomGenerator<int[]> generator = arrayArbitrary.generator(1);
+
+		Shrinkable<int[]> array = generator.next(random);
+		assertThat(array.value().length).isBetween(0, 5);
+		List<Integer> actual = IntStream.of(array.value()).boxed().collect(Collectors.toList());
+		assertThat(actual).isSubsetOf(1, 2, 3, 4, 5, 6, 7, 8, 9, 10);
+	}
+
+	@Example
+	void edgeCases() {
+		Arbitrary<Integer> ints = Arbitraries.of(-10, 10);
+		ArrayArbitrary<Integer, Integer[]> arbitrary = ints.array(Integer[].class);
+		assertThat(collectEdgeCases(arbitrary.edgeCases())).containsExactlyInAnyOrder(
+				new Integer[]{},
+				new Integer[]{-10},
+				new Integer[]{10}
+		);
+		assertThat(collectEdgeCases(arbitrary.edgeCases())).hasSize(3);
+	}
+
+	@Group
+	class ExhaustiveGeneration {
+
+		@Example
+		void arraysAreCombinationsOfElementsUpToMaxLength() {
+			Optional<ExhaustiveGenerator<Integer[]>> optionalGenerator =
+					Arbitraries.integers().between(1, 2).array(Integer[].class)
+							   .ofMaxSize(2).exhaustive();
+			assertThat(optionalGenerator).isPresent();
+
+			ExhaustiveGenerator<Integer[]> generator = optionalGenerator.get();
+			assertThat(generator.maxCount()).isEqualTo(7);
+			assertThat(generator).containsExactly(
+					new Integer[]{},
+					new Integer[]{1},
+					new Integer[]{2},
+					new Integer[]{1, 1},
+					new Integer[]{1, 2},
+					new Integer[]{2, 1},
+					new Integer[]{2, 2}
+			);
+		}
+
+		@Example
+		void elementArbitraryNotExhaustive() {
+			Optional<ExhaustiveGenerator<Double[]>> optionalGenerator =
+					Arbitraries.doubles().between(1, 10).array(Double[].class).ofMaxSize(1).exhaustive();
+			assertThat(optionalGenerator).isNotPresent();
+		}
+
+		@Example
+		void tooManyCombinations() {
+			Optional<ExhaustiveGenerator<Integer[]>> optionalGenerator =
+					Arbitraries.integers().between(1, 10).array(Integer[].class).ofMaxSize(10).exhaustive();
+			assertThat(optionalGenerator).isNotPresent();
+		}
+	}
+
+	@Group
+	@PropertyDefaults(tries = 100)
+	class Shrinking {
+
+		@Property
+		void shrinksToEmptyArrayByDefault(@ForAll Random random) {
+			ArrayArbitrary<Integer, Integer[]> arrays = Arbitraries.integers().between(1, 10).array(Integer[].class);
+			Integer[] value = falsifyThenShrink(arrays, random);
+			assertThat(value).isEmpty();
+		}
+
+		@Property
+		void shrinkToMinSize(@ForAll Random random, @ForAll @IntRange(min = 1, max = 20) int min) {
+			ArrayArbitrary<Integer, Integer[]> arrays = Arbitraries.integers().between(1, 10).array(Integer[].class).ofMinSize(min);
+			Integer[] value = falsifyThenShrink(arrays, random);
+			assertThat(value).hasSize(min);
+			assertThat(value).containsOnly(1);
+		}
+
+	}
+
+}